refactor(harness): extract /api/agent routes from server.create_app
Move the real-repo-run HTTP surface into harness/agent_routes.py using the
same register-and-inject pattern as gate_ops.py. create_app keeps the
guarded chain and generation/agent-run gates; route bodies no longer live
inside the 1400-line factory. Look up monkeypatched names on harness.server
at request time so existing tests keep working. No APIRouter (FastAPI 0.138
hides those from app.routes). Behavior unchanged.
